+++ packages/snap-preact-components/src/components/Atoms/Dropdown/Dropdown.test.tsx
@@ -48,16 +48,16 @@ describe('Dropdown Component', () => {
 		const contentText = 'this is the content';
 		const rendered = render(<Dropdown button={'open me'} content={contentText} />);
 
-		const buttonElement = rendered.getByText(contentText);
-		expect(buttonElement).toBeInTheDocument();
+		const contentElement = rendered.getByText(contentText);
+		expect(contentElement).toBeInTheDocument();
 	});
 
 	it('renders children prop', () => {
 		const child = 'this is the child';
 		const rendered = render(<Dropdown button={'open me'}>{child}</Dropdown>);
 
-		const buttonElement = rendered.getByText(child);
-		expect(buttonElement).toBeInTheDocument();
+		const childElement = rendered.getByText(child);
+		expect(childElement).toBeInTheDocument();
 	});
 
 	it('renders content and children props', () => {
@@ -71,6 +71,82 @@ describe('Dropdown Component', () => {
 
 		const buttonElement = rendered.getByText(contentText + child);
 		expect(buttonElement).toBeInTheDocument();
+	});
+
+	it('renders button prop with props', async () => {
+		const Button = ({ open }: { open?: boolean }) => {
+			return <div className="button-with-props">{open ? 'close' : 'open'} me</div>;
+		};
+		const rendered = render(<Dropdown button={<Button />}></Dropdown>);
+
+		const buttonElement = rendered.container.querySelector('.button-with-props');
+		expect(buttonElement).toBeInTheDocument();
+		expect(buttonElement.innerHTML).toBe('open me');
+
+		await (buttonElement as HTMLElement).click();
+		expect(buttonElement.innerHTML).toBe('close me');
+	});
+
+	it('renders children prop with props', async () => {
+		const Child = ({ open, toggleOpen }: { open?: boolean; toggleOpen?: () => void }) => {
+			return (
+				<div
+					className="child-with-props"
+					onClick={() => {
+						toggleOpen();
+					}}
+				>
+					im {open ? 'open' : 'closed'}
+				</div>
+			);
+		};
+		const rendered = render(
+			<Dropdown button={'open me'}>
+				<Child />
+			</Dropdown>
+		);
+
+		const buttonElement = rendered.getByText('open me');
+		expect(buttonElement).toBeInTheDocument();
+
+		const childElement = rendered.container.querySelector('.child-with-props');
+		expect(childElement).toBeInTheDocument();
+		expect(childElement.innerHTML).toBe('im closed');
+
+		await buttonElement.click();
+		expect(childElement.innerHTML).toBe('im open');
+
+		await (childElement as HTMLElement).click();
+		expect(childElement.innerHTML).toBe('im closed');
+	});
+
+	it('renders content prop with props', async () => {
+		const Content = ({ open, toggleOpen }: { open?: boolean; toggleOpen?: () => void }) => {
+			return (
+				<div
+					className="content-with-props"
+					onClick={() => {
+						toggleOpen();
+					}}
+				>
+					im {open ? 'open' : 'closed'}
+				</div>
+			);
+		};
+		const rendered = render(<Dropdown button={'open me'} content={<Content />}></Dropdown>);
+
+		const buttonElement = rendered.getByText('open me');
+		expect(buttonElement).toBeInTheDocument();
+
+		const contentElement = rendered.container.querySelector('.content-with-props');
+		expect(contentElement).toBeInTheDocument();
+		expect(contentElement.innerHTML).toBe('im closed');
+
+		await buttonElement.click();
+		expect(contentElement.innerHTML).toBe('im open');
+
+		await (contentElement as HTMLElement).click();
+		expect(contentElement.innerHTML).toBe('im closed');
 	});
 
 	it('disables styles', () => {
@@ -91,7 +167,7 @@ describe('Dropdown Component', () => {
 
 		const rendered = render(
 			<div>
-				<span class="outside">outside</span>
+				<span className="outside">outside</span>
 				<Dropdown content={contentText} onClick={clickFn} button={buttonText} onToggle={toggleFn} />
 			</div>
 		);
@@ -113,7 +189,7 @@ describe('Dropdown Component', () => {
 
 		const rendered = render(
 			<div>
-				<span class="outside">outside</span>
+				<span className="outside">outside</span>
 				<Dropdown disableClickOutside startOpen content={contentText} button={buttonText} onToggle={toggleFn} />
 			</div>
 		);